backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
httpx
